Combine commitlog test files
